lsm/db: move to multi-actor model
Currently, the LSM tree has one big background loop for all the work
that it does. However, logically these can be split into different
fibers. Flushing and compaction can never overlap with each other
so we can run them in parallel, and have a third fiber that linearizes
the manifest writes. Note there is opportunity to collapse multiple
manifest updates, but that will come as a future optimization.

N.B. There is a bug currently with the GC Actor. Can you spot it? :)
It will be fixed in the next commit.
