+++ src/v/lsm/db/tests/gc_actor_test.cc
@@ -96,7 +96,7 @@ TEST_F(GcActorTest, LiveFilesNotDeleted) {
     msg.live_files.insert(fh1);
     msg.live_files.insert(fh2);
     msg.live_files.insert(fh3);
-    _gc->tell(std::move(msg)).get();
+    _gc->tell(std::move(msg));
     tests::drain_task_queue().get();
 
     EXPECT_EQ(list_files().size(), 3);
@@ -116,7 +116,7 @@ TEST_F(GcActorTest, UnusedFilesDeletedImmediatelyWithZeroDelay) {
     msg.highest_used_file_id = 3_file_id;
     msg.live_files.insert(fh1);
     msg.live_files.insert(fh3);
-    _gc->tell(std::move(msg)).get();
+    _gc->tell(std::move(msg));
     tests::drain_task_queue().get();
 
     std::ignore = fh2; // deleted
@@ -135,7 +135,7 @@ TEST_F(GcActorTest, FutureEpochFilesNotDeleted) {
 
     lsm::db::gc_message msg;
     msg.highest_used_file_id = 2_file_id;
-    _gc->tell(std::move(msg)).get();
+    _gc->tell(std::move(msg));
     tests::drain_task_queue().get();
 
     std::ignore = fh_current; // deleted
@@ -159,7 +159,7 @@ TEST_F(GcActorTest, StopTrackingExternalDeletes) {
     msg.highest_used_file_id = 3_file_id;
     msg.live_files.insert(fh_2);
     msg.live_files.insert(fh_3);
-    _gc->tell(std::move(msg)).get();
+    _gc->tell(std::move(msg));
     tests::drain_task_queue().get();
 
     EXPECT_THAT(list_files(), testing::ElementsAre(fh_1, fh_2, fh_3));
@@ -171,7 +171,7 @@ TEST_F(GcActorTest, StopTrackingExternalDeletes) {
     msg = {};
     msg.highest_used_file_id = 3_file_id;
     msg.live_files.insert(fh_3);
-    _gc->tell(std::move(msg)).get();
+    _gc->tell(std::move(msg));
     tests::drain_task_queue().get();
 
     EXPECT_THAT(list_files(), testing::ElementsAre(fh_2, fh_3));
@@ -181,7 +181,7 @@ TEST_F(GcActorTest, StopTrackingExternalDeletes) {
     msg = {};
     msg.highest_used_file_id = 3_file_id;
     msg.live_files.insert(fh_3);
-    _gc->tell(std::move(msg)).get();
+    _gc->tell(std::move(msg));
     tests::drain_task_queue().get();
 
     EXPECT_THAT(list_files(), testing::ElementsAre(fh_3));
